+++ base/exports.jl
@@ -790,6 +790,7 @@ export
     parentmodule,
     pathof,
     pkgdir,
+    pkgversion,
     names,
     which,
     @isdefined,