+++ NEWS.md
@@ -67,6 +67,8 @@ New library functions
 * `Iterators.flatmap` was added ([#44792]).
 * New helper `Splat(f)` which acts like `x -> f(x...)`, with pretty printing for
   inspecting which function `f` was originally wrapped. ([#42717])
+* New `pkgversion(m::Module)` function to get the version of the package that loaded
+  a given module, similar to `pkgdir(m::Module)`. ([#45607])
 
 Library changes
 ---------------